+++ ProjectE-Integration/src/api/java/com/tagnumelite/projecteintegration/api/utils/Utils.java
@@ -1,21 +1,11 @@
 package com.tagnumelite.projecteintegration.api.utils;
 
 import com.tagnumelite.projecteintegration.api.PEIApi;
-import com.tagnumelite.projecteintegration.api.internal.sized.SizedIngredient;
-import moze_intel.projecte.emc.IngredientMap;
-import net.minecraft.block.Block;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.crafting.Ingredient;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TextComponentString;
 import net.minecraft.util.text.TextFormatting;
-import net.minecraftforge.fluids.FluidStack;
-import org.apache.commons.lang3.ClassUtils;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.List;
 
 public class Utils {
@@ -29,115 +19,6 @@ public class Utils {
     public static boolean isSameItem(ItemStack stack1, ItemStack stack2) {
         return stack2.getItem() == stack1.getItem()
             && (stack2.getItemDamage() == 32767 || stack2.getItemDamage() == stack1.getItemDamage());
-    }
-
-    public static IngredientMap<Object> createInputs(Object... inputs) {
-        return getInputsFirst(new IngredientMap<>(), inputs);
-    }
-
-    //TODO: Remove duplicated code, and look for performance improvements.
-    public static IngredientMap<Object> getInputsFirst(IngredientMap<Object> ingredients, Object... inputs) {
-        for (Object input : inputs) {
-            if (input == null)
-                continue;
-
-            if (input instanceof ItemStack) {
-                if (((ItemStack) input).isEmpty())
-                    continue;
-
-                ingredients.addIngredient(input, ((ItemStack) input).getCount());
-            } else if (input instanceof Item || input instanceof Block || input instanceof String
-                || input.getClass().equals(Object.class)) {
-                ingredients.addIngredient(input, 1);
-            } else if (input instanceof FluidStack) {
-                if (((FluidStack) input).amount <= 0)
-                    continue;
-
-                ingredients.addIngredient(input, ((FluidStack) input).amount);
-            } else if (input instanceof List) {
-                List<?> inputt = (List<?>) input;
-                if (inputt.isEmpty())
-                    continue;
-
-                if (inputt.size() == 1) {
-                    getInputsFirst(ingredients, inputt.get(0));
-                    continue;
-                }
-
-                ingredients.addIngredient(PEIApi.getList((List<?>) input), 1);
-            } else if (input instanceof Ingredient) {
-                if (input == Ingredient.EMPTY)
-                    continue;
-
-                ingredients.addIngredient(PEIApi.getIngredient((Ingredient) input), 1);
-            } else if (input instanceof SizedIngredient) {
-                SizedIngredient inp = (SizedIngredient) input;
-                if (inp.object == Ingredient.EMPTY)
-                    continue;
-
-                ingredients.addIngredient(PEIApi.getIngredient(inp.object), inp.amount);
-            } else if (input instanceof Object[]) {
-                PEIApi.LOG.debug("Found Array within Array: {} within {}", input, inputs);
-                getInputsSecond(ingredients, (Object[]) input);
-            } else {
-                PEIApi.LOG.warn("Unknown Input: {} ({})", input, ClassUtils.getPackageCanonicalName(input.getClass()));
-                continue;
-            }
-        }
-
-        return ingredients;
-    }
-
-    public static IngredientMap<Object> getInputsSecond(IngredientMap<Object> ingredients, Object... inputs) {
-        for (Object input : inputs) {
-            if (input == null)
-                continue;
-
-            if (input instanceof ItemStack) {
-                if (((ItemStack) input).isEmpty())
-                    continue;
-
-                ingredients.addIngredient(input, ((ItemStack) input).getCount());
-            } else if (input instanceof Item || input instanceof Block || input instanceof String
-                || input.getClass().equals(Object.class)) {
-                ingredients.addIngredient(input, 1);
-            } else if (input instanceof FluidStack) {
-                if (((FluidStack) input).amount <= 0)
-                    continue;
-
-                ingredients.addIngredient(input, ((FluidStack) input).amount);
-            } else if (input instanceof List) {
-                List<?> inputt = (List<?>) input;
-                if (inputt.isEmpty())
-                    continue;
-
-                if (inputt.size() == 1) {
-                    getInputsFirst(ingredients, inputt.get(0));
-                    continue;
-                }
-
-                ingredients.addIngredient(PEIApi.getList((List<?>) input), 1);
-            } else if (input instanceof Ingredient) {
-                if (input == Ingredient.EMPTY)
-                    continue;
-
-                ingredients.addIngredient(PEIApi.getIngredient((Ingredient) input), 1);
-            } else if (input instanceof SizedIngredient) {
-                SizedIngredient inp = (SizedIngredient) input;
-                if (inp.object == Ingredient.EMPTY)
-                    continue;
-
-                ingredients.addIngredient(PEIApi.getIngredient(inp.object), inp.amount);
-            } else if (input instanceof Object[]) {
-                PEIApi.LOG.debug("Found Array within Array: {} within {}", input, inputs);
-                getInputsFirst(ingredients, Arrays.asList(input));
-            } else {
-                PEIApi.LOG.warn("Unknown Input: {} ({})", input, ClassUtils.getPackageCanonicalName(input.getClass()));
-                continue;
-            }
-        }
-
-        return ingredients;
     }
 
     /**